requirements.txt
django==4.1
djangorestframework==3.13.0
drf-yasg==1.21
djangorestframework-simplejwt==5.2.2

